+++ src/main/java/ch/heig/statique/Statique.java
@@ -3,6 +3,8 @@ package ch.heig.statique;
 import ch.heig.statique.Commands.*;
 import picocli.CommandLine;
 
+import java.io.IOException;
+import java.util.Properties;
 import java.util.concurrent.Callable;
 
 @CommandLine.Command(
@@ -13,8 +15,13 @@ import java.util.concurrent.Callable;
                 Clean.class,
                 Build.class,
                 Serve.class,
-        })
+        },
+        versionProvider = VersionProvider.class)
 public class Statique implements Callable<Integer> {
+
+    @CommandLine.Option(names = { "-V", "--version" }, versionHelp = true,
+            description = "Print version information and exit")
+    boolean versionRequested;
 
     public static void main( String[] args ) {
         int exitCode = new CommandLine(new Statique()).execute(args);
@@ -27,5 +34,16 @@ public class Statique implements Callable<Integer> {
     public Integer call() {
         CommandLine.usage(this, System.out);
         return 0;
+    }
+}
+
+class VersionProvider implements CommandLine.IVersionProvider {
+    public String[] getVersion() throws IOException {
+
+        final Properties properties = new Properties();
+        properties.load(this.getClass().getClassLoader().getResourceAsStream("project.properties"));
+        return new String[] {
+                properties.getProperty("name") + " v" + properties.getProperty("version")
+        };
     }
 }
